+++ lib/src/helper/baidu_upload_helper.dart
@@ -1,4 +1,3 @@
-import 'dart:collection';
 import 'dart:convert';
 import 'dart:io';
 
@@ -67,10 +66,8 @@ class BaiduUploadHelper with ILogger {
   final Map<int, String> _blockMd5Map = {};
 
   /// 已上传的文件块
-  Map<int, String> get uploadedBlocks {
-    return LinkedHashMap.fromEntries(
-      _blockMd5Map.entries.toList()..sort((a, b) => a.key - b.key),
-    );
+  List<_PartBlock> get uploadedBlocks {
+    return _blockMd5Map.entries.map((e) => _PartBlock(e.key, e.value)).toList();
   }
 
   bool _isUploading = false;
@@ -80,6 +77,12 @@ class BaiduUploadHelper with ILogger {
 
   String? _uploadId;
 
+  /// 共有多少个文件块
+  int totalBlockCount = 0;
+
+  /// uploadCount 已上传的文件块数量
+  int uploadCount = 0;
+
   /// 获取应该被保存的 map
   Map<String, dynamic> saveProgress() {
     if (_uploadId == null) {
@@ -88,7 +91,7 @@ class BaiduUploadHelper with ILogger {
 
     return {
       'uploadId': _uploadId,
-      'uploadedBlocks': uploadedBlocks,
+      'uploadedBlocks': uploadedBlocks.map((e) => e.toJson()).toList(),
       'localPath': localPath,
       'remotePath': remotePath,
       'memberLevel': memberLevel,
@@ -110,7 +113,12 @@ class BaiduUploadHelper with ILogger {
     }
 
     _uploadId = progress['uploadId'];
-    _blockMd5Map.addAll(progress['uploadedBlocks']);
+
+    final uploadedBlocks = progress['uploadedBlocks'] as List;
+    uploadedBlocks.forEach((e) {
+      final block = _PartBlock.fromMap(e);
+      _blockMd5Map[block.block] = block.md5;
+    });
   }
 
   /// 保存进度到文件
@@ -166,6 +174,8 @@ class BaiduUploadHelper with ILogger {
     final uploadId = preCreate.uploadId;
     _uploadId = uploadId;
 
+    totalBlockCount = preCreate.blockList.length;
+
     for (final blockIndex in preCreate.blockList) {
       if (_blockMd5Map.containsKey(blockIndex)) {
         log('已经上传过的块：$blockIndex, 跳过');
@@ -180,6 +190,7 @@ class BaiduUploadHelper with ILogger {
       );
 
       _blockMd5Map[blockIndex] = block.md5;
+      uploadCount++;
       uploadHandler?.onUploadPartComplete(this, blockIndex, block.md5);
     }
 
@@ -228,4 +239,22 @@ mixin UploadHelperListener {
     Object error,
     StackTrace stackTrace,
   ) {}
+}
+
+class _PartBlock {
+  final int block;
+  final String md5;
+
+  _PartBlock(this.block, this.md5);
+
+  _PartBlock.fromMap(Map map)
+      : block = map['block'] as int,
+        md5 = map['md5'] as String;
+
+  Map<String, dynamic> toJson() {
+    return {
+      'block': block,
+      'md5': md5,
+    };
+  }
 }