+++ publish.sh
@@ -1,0 +1,4 @@
+git push origin
+git push github
+unset PUB_HOSTED_URL
+dart pub publish