+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## 1.0.4
+
+- 修复了一处 `BaiduUploadHelper` 恢复进度的 bug
+
+## 1.0.3
+
+- 修复 bug： `BaiduUploadHelper`
+- 为 `BaiduUploadHelper` 添加 `uploadCount` 来标识本地上传的数量
+
+## 1.0.2
+
+- 为 `BaiduUploadHelper` 添加了一个 `totalBlockCount` 用来标识文件块的数量
+
 ## 1.0.1
 
 - 清理了代码